[Dagger]: Fix binds method validation caching.
AnyBindingMethodValidator is used by both ModuleProcessingStep and BindsMethodProcessingStep, which can lead to duplicate error reports for the same binding method. BindsMethodProcessingStep does check if a method has already been validated, but it doesn't actually work since AnyBindingMethodValidator is not scoped.

This CL scopes the reports in AnyBindingMethodValidator to prevent reporting duplicate errors.

RELNOTES=N/A
PiperOrigin-RevId: 335063565